04-Cloud_Computing-Managing_Data-Environmental_Data/NetCDF2Cloud/requirements.txt
NetCDF4
numpy
apache-libcloud
boto3